added function write_spherical_tensor() in formal.c
this function runs at the end of RH and calculates the anisotropy in the radiation field. It uses the short characterstics solution in formal.c to get the rays. and then I do a wavlength, angle intergration to calculate components of spherical tensors.

Note that the wavelength integration is rather crude since I've hard coded in the wavlengths of the strontium line where I'm integrating.
Finally, it outputs these arrays at txt files.